+++ realm/realm-library/src/main/java/io/realm/RealmObject.java
@@ -244,6 +244,12 @@ public abstract class RealmObject implements RealmModel {
             if (!listeners.contains(listener)) {
                 listeners.add(listener);
             }
+            if (isLoaded(proxy)) {
+                // Try to add this object to the realmObjects if it has already been loaded.
+                // For newly created async objects, it will be handled in RealmQuery.findFirstAsync &
+                // HandlerController.completedAsyncRealmObject.
+                realm.handlerController.addToRealmObjects(proxy);
+            }
         } else {
             throw new IllegalArgumentException("Cannot add listener from this unmanaged RealmObject (created outside of Realm)");
         }